disable avx_memcpy for now, open avi file using mmap (default to 1% of total memory per AVI, set with VEEJAY_MMAP_PER_FILE, drop map-from-file commandline parameter, add mmap to AVI open

+++ veejay-current/veejay-server/libel/elcache.c
@@ -147,7 +147,7 @@ void	reset_cache(void *cache)
 	}
 }
 
-int	cache_avail_mb()
+long	cache_avail_mb()
 {
 	return ( total_mem_used_ == 0 ? 0 : total_mem_used_ / (1024 * 1024 ));
 }
